Get rid of PreviousDepGraph.
